Fix integer type of user and group ids
User IDs are INTEGER in database. The java classes are fixed
to replace long by int.

+++ server/sonar-db-migration/src/main/java/org/sonar/server/platform/db/migration/version/v62/PopulateOrganizationUuidOfGroups.java
@@ -49,10 +49,10 @@ public class PopulateOrganizationUuidOfGroups extends DataChange {
     massUpdate.update("update groups set organization_uuid=?, updated_at=? where id=?");
     massUpdate.rowPluralName("groups");
     massUpdate.execute((row, update) -> {
-      long groupId = row.getLong(1);
+      int groupId = row.getInt(1);
       update.setString(1, organizationUuid);
       update.setDate(2, new Date(system2.now()));
-      update.setLong(3, groupId);
+      update.setInt(3, groupId);
       return true;
     });
   }
